Move escalation tooltip to tooltips folder
